[WIP] Add Rotate UI #112
