+++ packages/generator/package.json
@@ -34,7 +34,7 @@
     "node": ">=14"
   },
   "scripts": {
-    "develop": "tsc -w",
+    "develop": "tsc -w -p tsconfig.esm.json",
     "build": "npm run build:cjs && npm run build:esm",
     "build:cjs": "tsc -p tsconfig.cjs.json",
     "build:esm": "tsc -p tsconfig.esm.json",
