+++ packages/ui/src/components/Schemas/SchemaUI.tsx
@@ -37,6 +37,7 @@ const Wrapper = ({
       width: schema.width * ZOOM,
       top: schema.position.y * ZOOM,
       left: schema.position.x * ZOOM,
+      transform: `rotate(${schema.rotate || 0}deg)`,
       border,
     }}
   >
